Fix non working (leaders) example
`/api/leaders` resource does no longer exist. Though it is possible to get a list
of users ordered by their "knowledge" rank. Also the "avatar" field is stored
in user's model already.

The other problem was connected with the file extension, it was always `png`,
but stepik uses `svg` actively too.

The last point is to simplify understanding of the code — just write `response.content`,
there is no need in `shutil.copyfileobj`.